+++ include/assert.h
@@ -4,7 +4,13 @@
 #ifdef NDEBUG
 #define assert(expr) ((void)0)
 #else
+#ifdef __cplusplus
+extern "C" {
+#endif
 void abort(void) __attribute__((noreturn));
+#ifdef __cplusplus
+}
+#endif
 #define assert(expr) ((expr) ? (void)0 : abort())
 #endif